feat(hisRootCh): historic root challenge

+++ nightfall-deployer/contracts/Utils.sol
@@ -26,27 +26,37 @@ library Utils {
     );
   }
 
-  function hashBlock(Structures.Block memory b, Structures.Transaction[] memory t) internal pure returns(bytes32) {
+  function hashBlock(Structures.Block memory b, Structures.Transaction[] memory t)
+  internal
+  pure
+  returns (bytes32)
+  {
     return keccak256(abi.encode(b, t));
   }
 
-  function compressG1(uint x, uint y) internal pure returns(uint) {
+  function compressG1(uint256 x, uint256 y) internal pure returns (uint256) {
     // compute whether y is odd or even and multiply by 2**255
-    uint parity = 0x8000000000000000000000000000000000000000000000000000000000000000 * (y % 2);
+    uint256 parity =
+    0x8000000000000000000000000000000000000000000000000000000000000000 * (y % 2);
     // add the parity bit to the x cordinate (x,y are 254 bits long - the final
     // number is 256 bits to fit with an Ethereum word)
     return parity + x;
   }
 
-  function compressG2(uint xr, uint xi, uint yr, uint yi) internal pure returns(uint, uint) {
+  function compressG2(
+    uint256 xr,
+    uint256 xi,
+    uint256 yr,
+    uint256 yi
+  ) internal pure returns (uint256, uint256) {
     return (compressG1(xr, yr), compressG1(xi, yi));
   }
 
-  function compressProof(uint[8] memory proof) internal pure returns(uint[4] memory) {
-    uint a = compressG1(proof[0], proof[1]);
-    (uint rb, uint ib) = compressG2(proof[2], proof[3], proof[4], proof[5]);
-    uint c = compressG1(proof[6], proof[7]);
-    return [ a, rb, ib, c];
+  function compressProof(uint256[8] memory proof) internal pure returns (uint256[4] memory) {
+    uint256 a = compressG1(proof[0], proof[1]);
+    (uint256 rb, uint256 ib) = compressG2(proof[2], proof[3], proof[4], proof[5]);
+    uint256 c = compressG1(proof[6], proof[7]);
+    return [a, rb, ib, c];
   }
 
   // counts the number of non-zero values (useful for counting real commitments and nullifiers)
